requirements.txt
paramiko>=2.12.0
pyyaml>=6.0
pytest>=7.4.0
openpyxl>=3.1.0

